# backend/app/api/v1/endpoints/executions.py
"""
执行管理端点

测试执行的启动、监控和结果查询
"""

from typing import List, Optional
from fastapi import APIRouter, HTTPException, Query, BackgroundTasks
from pydantic import BaseModel, Field

router = APIRouter()


# ===================== 数据模型 =====================

class ExecutionRequest(BaseModel):
    """执行请求"""
    tc_ids: List[str] = Field(..., description="测试用例 ID 列表")
    mode: str = Field("normal", description="执行模式: normal, debug, fast")
    parallel: bool = Field(True, description="是否并行执行")
    max_workers: int = Field(5, ge=1, le=20, description="最大并行数")
    env: Optional[str] = Field(None, description="目标环境")


class ExecutionResponse(BaseModel):
    """执行响应"""
    execution_id: str
    status: str
    total_cases: int
    dashboard_url: str


class ExecutionStatus(BaseModel):
    """执行状态"""
    execution_id: str
    status: str
    progress: float
    passed: int
    failed: int
    skipped: int
    running: int
    pending: int
    start_time: str
    elapsed_seconds: float


class ExecutionResult(BaseModel):
    """执行结果"""
    execution_id: str
    status: str
    summary: dict
    cases: List[dict]
    duration_seconds: float
    report_url: Optional[str]


# ===================== API 端点 =====================

@router.post("", response_model=ExecutionResponse, status_code=202)
async def start_execution(
    request: ExecutionRequest,
    background_tasks: BackgroundTasks,
):
    """
    启动测试执行
    
    将 TC-IR 分发到执行引擎，返回执行 ID
    """
    # TODO: 创建执行任务，发送到 Celery
    return ExecutionResponse(
        execution_id="EXEC_PLACEHOLDER",
        status="pending",
        total_cases=len(request.tc_ids),
        dashboard_url="/executions/EXEC_PLACEHOLDER",
    )


@router.get("/{execution_id}", response_model=ExecutionStatus)
async def get_execution_status(execution_id: str):
    """
    获取执行状态
    
    实时查询执行进度
    """
    # TODO: 从 Redis/数据库查询状态
    raise HTTPException(status_code=404, detail=f"执行 {execution_id} 不存在")


@router.get("/{execution_id}/result", response_model=ExecutionResult)
async def get_execution_result(execution_id: str):
    """
    获取执行结果
    
    返回完整的执行报告
    """
    # TODO: 从数据库查询结果
    raise HTTPException(status_code=404, detail=f"执行 {execution_id} 不存在")


@router.post("/{execution_id}/cancel", status_code=202)
async def cancel_execution(execution_id: str):
    """
    取消执行
    
    终止正在运行的执行任务
    """
    # TODO: 发送取消信号到 Celery
    return {"message": f"取消请求已发送: {execution_id}"}


@router.get("", response_model=List[ExecutionStatus])
async def list_executions(
    status: Optional[str] = Query(None, description="状态过滤"),
    limit: int = Query(20, ge=1, le=100, description="返回数量"),
):
    """
    获取执行列表
    
    查询最近的执行记录
    """
    # TODO: 从数据库查询
    return []