# backend/app/api/v1/endpoints/phoenix.py
"""
凤凰涅槃层端点

Trace → 脚本编译
"""

from typing import Optional
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel, Field

router = APIRouter()


# ===================== 数据模型 =====================

class CompileRequest(BaseModel):
    """编译请求"""
    trace_id: str = Field(..., description="Trace ID")
    strategy: str = Field("optimized", description="编译策略: exact, optimized, data_driven, hybrid")
    output_format: str = Field("pytest", description="输出格式: pytest, unittest")
    options: dict = Field(default_factory=dict, description="编译选项")


class CompileResponse(BaseModel):
    """编译响应"""
    script_id: str
    file_path: str
    code_preview: str
    parameters_extracted: int
    steps_generated: int


class ScriptInfo(BaseModel):
    """脚本信息"""
    script_id: str
    file_path: str
    name: str
    source_tc_id: str
    strategy: str
    created_at: str
    updated_at: str


class VersionInfo(BaseModel):
    """版本信息"""
    version: str
    author: str
    date: str
    message: str
    changes: str


# ===================== API 端点 =====================

@router.post("/compile", response_model=CompileResponse)
async def compile_trace(request: CompileRequest):
    """
    编译执行轨迹为测试脚本
    
    将 Trace Log 转化为 Python 测试代码
    """
    # TODO: 调用凤凰涅槃层
    raise HTTPException(status_code=501, detail="尚未实现")


@router.get("/scripts/{script_id}", response_model=ScriptInfo)
async def get_script(script_id: str):
    """
    获取脚本信息
    """
    # TODO: 从资产库查询
    raise HTTPException(status_code=404, detail=f"脚本 {script_id} 不存在")


@router.get("/scripts/{script_id}/code")
async def get_script_code(script_id: str):
    """
    获取脚本代码内容
    """
    # TODO: 从文件系统读取
    raise HTTPException(status_code=404, detail=f"脚本 {script_id} 不存在")


@router.get("/scripts/{script_id}/history")
async def get_script_history(
    script_id: str,
    limit: int = 20,
):
    """
    获取脚本版本历史
    """
    # TODO: 从 Git 查询
    return []


@router.post("/scripts/{script_id}/heal")
async def apply_healing(
    script_id: str,
    healing_record_id: str,
):
    """
    应用自愈修复
    
    将成功的修复策略固化到脚本
    """
    # TODO: 调用自愈编译
    raise HTTPException(status_code=501, detail="尚未实现")