# deploy/openapi.yaml
openapi: 3.0.3
info:
  title: 重明 (ChongMing) API
  description: |
    重明 AI 自动化测试平台 API 文档
    
    ## 核心模块
    - **神经设计层**: PRD 解析、用例生成
    - **左瞳引擎**: API 测试执行
    - **右瞳引擎**: UI 测试执行
    - **涡轮引擎**: 性能测试
    - **凤凰涅槃层**: 脚本编译
    - **缺陷分析**: 根因分析、修复建议
    
  version: 2.0.0
  contact:
    name: ChongMing Team
    email: chongming@example.com
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8000/api/v1
    description: 开发环境
  - url: https://test-api.chongming.ai/api/v1
    description: 测试环境
  - url: https://api.chongming.ai/api/v1
    description: 生产环境
...
paths:
...
  /executions:
    post:
      tags: [executions]
      summary: 创建执行任务
      operationId: createExecution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tc_ids]
              properties:
                tc_ids:
                  type: array
                  items:
                    type: string
                environment:
                  type: string
                  default: test
                parallel:
                  type: integer
                  default: 1
                config:
                  $ref: '#/components/schemas/ExecutionConfig'
      responses:
        '202':
          description: 执行任务已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, running]
  /executions/{execution_id}:
    get:
      tags: [executions]
      summary: 获取执行状态
      operationId: getExecution
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
...
  /api-engine/execute:
    post:
      tags: [api-engine]
      summary: 执行 API 测试
      operationId: executeApiTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_ir_chain]
              properties:
                api_ir_chain:
                  type: array
                  items:
                    $ref: '#/components/schemas/APIIR'
                environment:
                  type: string
      responses:
        '202':
          description: 执行任务已创建
  /api-engine/swagger/parse:
    post:
      tags: [api-engine]
      summary: 解析 Swagger 文档
      operationId: parseSwagger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                swagger_url:
                  type: string
                  format: uri
                swagger_content:
                  type: string
      responses:
        '200':
          description: 解析成功

  # ==================== 涡轮引擎 (性能测试) ====================
...
  /phoenix/compile:
    post:
      tags: [phoenix]
      summary: 编译执行轨迹为脚本
      operationId: compileTrace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trace_id]
              properties:
                trace_id:
                  type: string
                strategy:
                  type: string
                  enum: [simple, optimized, data_driven]
                  default: optimized
                framework:
                  type: string
                  enum: [pytest, unittest]
                  default: pytest
      responses:
        '200':
          description: 编译成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompiledScript'
...
  /phoenix/scripts/{script_id}/history:
    get:
      tags: [phoenix]
      summary: 获取脚本版本历史
      operationId: getScriptHistory
      parameters:
        - name: script_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功

  # ==================== 缺陷分析 ====================
...
components:
  schemas:
...
    APIIR:
      type: object
      description: API 中间表示
      properties:
        protocol:
          type: string
          default: API-IR
        version:
          type: string
          default: "1.0"
        id:
          type: string
        name:
          type: string
        request:
          type: object
          properties:
            method:
              type: string
              enum: [GET, POST, PUT, DELETE, PATCH]
            url:
              type: string
            headers:
              type: object
            body:
              type: object
        extraction:
          type: object
        assertion:
          type: object
    ExecutionConfig:
      type: object
      properties:
        browser:
          type: string
          default: chromium
        headless:
          type: boolean
          default: true
        timeout:
          type: integer
          default: 30000
        retries:
          type: integer
          default: 3
        enable_healing:
          type: boolean
          default: true
    ExecutionResult:
      type: object
      properties:
        execution_id:
          type: string
        status:
          type: string
          enum: [pending, running, passed, failed, cancelled]
        progress:
          type: number
        total_cases:
          type: integer
        passed:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
...
    CompiledScript:
      type: object
      properties:
        script_id:
          type: string
        tc_id:
          type: string
        content:
          type: string
        file_path:
          type: string
        framework:
          type: string
        parameters:
          type: array
          items:
            type: string
        git_commit:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []